Update Product model and refine Product Details UI
- Add description and rating fields to Product entity and UI state
- Refactor `ProductImageCarousel` with improved state management and layout
- Update Shopify configuration and build fields
- Improve product mapping logic for sizes, colors, and images
- Enhance `ProductDetailsTopBar` with click handling and UI state updates

+++ presintation/src/commonMain/kotlin/com/troves/presintation/ui/productDetails/ProductDetailUiState.kt
@@ -7,7 +7,7 @@ import com.troves.presintation.ui.productDetails.models.SizeUi
 data class ProductDetailUiState(
     val title: String = "",
     val priceFormatted: String= "",
-    val rating: Float = 0f,
+    val rating: Int = 0,
     val reviewCount: Int = 0,
     val description: String= "",
     val images: List<String> = emptyList(),
